requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-decouple
jinja2
aiosmtplib
email-validator
python-dotenv
pytz